requirements.txt
google-generativeai
httpx
python-dotenv
fastapi
uvicorn
selectolax
chromadb  # Vector database
sentence-transformers  # For embeddings